sysbench/pyproject.toml
[project]
name = "sysbench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "psycopg[binary]>=3.3.3",
    "pyyaml>=6.0.3",
]
